requirements.txt
# Bleeding edge Django
Django==1.8.4
django-appconf==1.0.1
django-bower==5.0.4
django-compressor==1.5
django-libsass==0.3
django-modelcluster==0.6.2
django-sweetcaptcha==0.1
django-taggit==0.17.1
django-treebeard==3.0

# REST
djangorestframework==3.2.3

# Models
django-model-utils==2.3.1

pytz==2015.4
six==1.9.0
wheel==0.24.0

# For the persistance stores
psycopg2==2.5.4

python-memcached==1.57
django-redis==4.2.0
elasticsearch==0.4.5

gunicorn==19.1.1

beautifulsoup4==4.4.0

html5lib==0.999
libsass==0.8.3

Pillow==2.9.0
Willow==0.2.1

requests==2.7.0
Unidecode==0.4.18

wagtail==1.1
wagtailsweetcaptcha==0.1.0

